refactor: move object domain to util
