+++ src/main/kotlin/util/Separator.kt
@@ -1,4 +1,4 @@
-package domain
+package util
 
 object Separator {
