+++ src/test/kotlin/util/SeparatorTest.kt
@@ -1,4 +1,4 @@
-package domain
+package util
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
